+++ animations/tools/config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  CDN_OBJECTS: [
+    'lottie',
+    'CHANGELOG.md',
+    'package.json',
+  ],
+};