+++ web-components/src/components/tabs/Tabs.ts
@@ -6,22 +6,22 @@
  *
  */
 
-import { html, internalProperty, LitElement, property, PropertyValues, query, queryAll } from "lit-element";
-import Sortable from "sortablejs";
-import reset from "@/wc_scss/reset.scss";
-import styles from "./scss/module.scss";
-import { customElementWithCheck, ResizeMixin, RovingTabIndexMixin, SlottedMixin } from "@/mixins";
-import { Tab, TabClickEvent } from "./Tab";
-import { TabPanel } from "./TabPanel";
-import { nanoid } from "nanoid";
-import { classMap } from "lit-html/directives/class-map";
-import { MenuOverlay } from "../menu-overlay/MenuOverlay"; // Keep type import as a relative path
-import { unsafeHTML } from "lit-html/directives/unsafe-html";
-import { repeat } from "lit-html/directives/repeat";
-import { styleMap } from "lit-html/directives/style-map";
 import "@/components/icon/Icon";
 import "@/components/menu-overlay/MenuOverlay";
 import { Key } from "@/constants";
+import { customElementWithCheck, ResizeMixin, RovingTabIndexMixin, SlottedMixin } from "@/mixins";
+import reset from "@/wc_scss/reset.scss";
+import { html, internalProperty, LitElement, property, PropertyValues, query, queryAll } from "lit-element";
+import { classMap } from "lit-html/directives/class-map";
+import { repeat } from "lit-html/directives/repeat";
+import { styleMap } from "lit-html/directives/style-map";
+import { unsafeHTML } from "lit-html/directives/unsafe-html";
+import { nanoid } from "nanoid";
+import Sortable from "sortablejs";
+import { MenuOverlay } from "../menu-overlay/MenuOverlay"; // Keep type import as a relative path
+import styles from "./scss/module.scss";
+import { Tab, TabClickEvent, TAB_CROSS_WIDTH } from "./Tab";
+import { TabPanel } from "./TabPanel";
 
 const MORE_MENU_TAB_TRIGGER_ID = "tab-more";
 const MORE_MENU_WIDTH = "264px"; // Designed width
@@ -52,7 +52,7 @@ export namespace Tabs {
     @property({ type: String, attribute: "chosen-class" }) chosenClass = "";
     @property({ type: Boolean, attribute: "force-fallback" }) forceFallback = false;
     @property({ type: String, attribute: "fallback-class" }) fallbackClass = "";
-    
+
     @internalProperty() private isMoreTabMenuVisible = false;
     @internalProperty() private isMoreTabMenuMeasured = false;
     @internalProperty() private isMoreTabMenuOpen = false;
@@ -68,6 +68,7 @@ export namespace Tabs {
     @query("slot[name='tab']") tabSlotElement!: HTMLSlotElement;
     @query("slot[name='panel']") panelSlotElement?: HTMLSlotElement;
     @query(".md-tab__list[part='tabs-list']") tabsListElement?: HTMLDivElement;
+    @query(".md-tabs__settings[part='md-tabs__settings']") tabsSettingsElement?: HTMLElement;
     @query(".md-menu-overlay__more_tab") moreTabMenuElement?: Tab.ELEMENT;
     @query("md-menu-overlay") menuOverlayElement?: MenuOverlay.ELEMENT;
 
@@ -82,10 +83,10 @@ export namespace Tabs {
         group: "shared",
         animation: 10,
         swapThreshold: 1,
-        draggable: "md-tab",  
+        draggable: "md-tab",
         direction: this.direction,
-        forceFallback: this.forceFallback,  
-        fallbackClass: this.fallbackClass,    
+        forceFallback: this.forceFallback,
+        fallbackClass: this.fallbackClass,
         ghostClass: this.ghostClass,
         chosenClass: this.chosenClass,
         onEnd: this.handleOnDragEnd
@@ -141,10 +142,12 @@ export namespace Tabs {
     // This operation may affect render performance when using frequently. Use careful!
     private measureTabsOffsetWidth() {
       return !this.justified && this.direction !== "vertical"
-        ? this.tabs.map((tab, idx) => tab.offsetWidth)
+        ? this.tabs.map((tab, idx) => {
+            return tab.closable ? tab.offsetWidth + TAB_CROSS_WIDTH : tab.offsetWidth;
+          })
         : this.tabs.map((tab, idx) => {
             tab.setAttribute("measuringrealwidth", "");
-            const offsetWidth = tab.offsetWidth;
+            const offsetWidth = tab.closable ? tab.offsetWidth + TAB_CROSS_WIDTH : tab.offsetWidth;
             tab.removeAttribute("measuringrealwidth");
             return offsetWidth;
           });
@@ -157,15 +160,30 @@ export namespace Tabs {
     private async manageOverflow() {
       if (this.direction !== "vertical") {
         let tabList;
-        if (this.tabsFilteredAsVisibleList.length === 0 && this.tabsFilteredAsHiddenList.length === 0)
+        if (this.tabsFilteredAsVisibleList.length === 0 && this.tabsFilteredAsHiddenList.length === 0) {
           tabList = [...this.tabs];
-        else tabList = [...this.tabsFilteredAsVisibleList, ...this.tabsFilteredAsHiddenList];
+        } else {
+          tabList = [...this.tabsFilteredAsVisibleList, ...this.tabsFilteredAsHiddenList];
+        }
+
+        tabList.forEach(tab => {
+          if (tab.children?.length && tab.children[0]?.children?.length === 0) {
+            const slotHeaderNode = tab
+              ?.querySelector("slot")
+              ?.assignedNodes({ flatten: true })[0]
+              .cloneNode(true);
+            if (slotHeaderNode) {
+              (slotHeaderNode as HTMLElement).classList.add("tab-content");
+              tab?.children[0]?.appendChild(slotHeaderNode);
+            }
+          }
+        });
 
         const tabsCount = tabList.length;
-        if (this.tabsListElement && tabsCount > 1) {
-          const tabsListViewportOffsetWidth = this.tabsListElement.offsetWidth;
-
-          // Awaiting all tabs updates
+        if (this.tabsListElement && tabsCount) {
+          const tabsListViewportOffsetWidth = this.tabsSettingsElement?.offsetWidth
+            ? this.tabsListElement.offsetWidth - this.tabsSettingsElement?.offsetWidth
+            : this.tabsListElement.offsetWidth;
           await this.ensureTabsUpdateComplete(this.tabs);
 
           const tabsOffsetsWidths = this.measureTabsOffsetWidth();
@@ -263,7 +281,7 @@ export namespace Tabs {
       }
 
       tabs.forEach((tab, index) => {
-        const id = "tab_" + nanoid(10); 
+        const id = "tab_" + nanoid(10);
         tab.setAttribute("id", id);
         tab.setAttribute("aria-controls", id);
         tab.selected = this.selected === index;
@@ -644,6 +662,7 @@ export namespace Tabs {
           if (isMoreTriggerTab) {
             //
           } else if (isVisibleTab) {
+            event.stopPropagation();
             this.changeSelectedTabIdx(this.selected === firstVisibleTabIdx ? lastVisibleTabIdx : this.selected - 1);
           } else if (isHiddenTab) {
             //
@@ -654,6 +673,7 @@ export namespace Tabs {
           if (isMoreTriggerTab) {
             //
           } else if (isVisibleTab) {
+            event.stopPropagation();
             this.changeSelectedTabIdx(this.selected === lastVisibleTabIdx ? firstVisibleTabIdx : this.selected + 1);
           } else if (isHiddenTab) {
             //
@@ -938,6 +958,9 @@ export namespace Tabs {
               )}
             </div>
           </md-menu-overlay>
+          <div class="md-tabs__settings" part="md-tabs__settings">
+            <slot name="settings"></slot>
+          </div>
         </div>
         <div
           part="tabs-content"