+++ illustrations/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/illustrations@1.5.0...@momentum-ui/illustrations@1.6.0) (2021-06-24)
+
+
+### Features
+
+* **illustrations:** add illustrations ([#958](https://github.com/momentum-design/momentum-ui/issues/958)) ([6044818](https://github.com/momentum-design/momentum-ui/commit/60448186d9e8afa22a9d824212e14d1d94c3bb0d))
+* **illustrations:** add illustrations ([#963](https://github.com/momentum-design/momentum-ui/issues/963)) ([c54844b](https://github.com/momentum-design/momentum-ui/commit/c54844b62d6a3b09d0cc86fe664c75dc3a145bab))
+
+
+
+
+
+# [1.5.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/illustrations@1.4.0...@momentum-ui/illustrations@1.5.0) (2021-06-22)
+
+
+### Features
+
+* **illustrations:** add illustrations ([#954](https://github.com/momentum-design/momentum-ui/issues/954)) ([0d5fcc3](https://github.com/momentum-design/momentum-ui/commit/0d5fcc3fc1e21c2125cc4ecf5abe8e871a931d70))
+* **illustrations:** add illustrations ([#956](https://github.com/momentum-design/momentum-ui/issues/956)) ([3177194](https://github.com/momentum-design/momentum-ui/commit/317719475a0780b23d7a117641035856c5dbd38a))
+
+
+
+
+
+# [1.4.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/illustrations@1.3.0...@momentum-ui/illustrations@1.4.0) (2021-06-18)
+
+
+### Features
+
+* **illustrations:** add illustrations ([#953](https://github.com/momentum-design/momentum-ui/issues/953)) ([9a90767](https://github.com/momentum-design/momentum-ui/commit/9a90767e24f25690b722dcb676758d57552f2ee6))
+
+
+
+
+
+# [1.3.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/illustrations@1.2.1...@momentum-ui/illustrations@1.3.0) (2021-06-17)
+
+
+### Features
+
+* **illustrations:** add illustrations ([#948](https://github.com/momentum-design/momentum-ui/issues/948)) ([3d8831d](https://github.com/momentum-design/momentum-ui/commit/3d8831d0cbe2dc47c8a95b7718e3a4417cb08333))
+* **illustrations:** add illustrations ([#949](https://github.com/momentum-design/momentum-ui/issues/949)) ([4a2b407](https://github.com/momentum-design/momentum-ui/commit/4a2b4079cd3fa520ca9f10b5b72497a06a2fc9e6))
+* **illustrations:** remove all existing illustrations ([#950](https://github.com/momentum-design/momentum-ui/issues/950)) ([579e096](https://github.com/momentum-design/momentum-ui/commit/579e0968327068b8a2c7f91c161e24e682f2f64a))
+
+
+
+
+
 ## [1.2.1](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/illustrations@1.2.0...@momentum-ui/illustrations@1.2.1) (2021-06-10)