+++ tokens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/tokens@1.7.0...@momentum-ui/tokens@1.7.1) (2021-06-29)
+
+**Note:** Version bump only for package @momentum-ui/tokens
+
+
+
+
+
 # [1.7.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/tokens@1.6.2...@momentum-ui/tokens@1.7.0) (2021-05-27)